third_party/WebKit/LayoutTests/external/wpt/cookies/http-state/0001.html

<!doctype html>
<html>
<head>
  <meta charset=utf-8>
  <title>Tests basic cookie setting functionality</title>
  <meta name=help href="https://tools.ietf.org/html/rfc6265#page-8">

  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>

  setup({ explicit_timeout: true });

  function delete_all_cookies(doc) {
    doc.cookie.split(";").forEach((c) => {
      doc.cookie = c.replace(/^ +/, "")
                    .replace(/=.*/, "=;expires=" +
                             new Date().toUTCString() +
                             ";path=/"); });
  }

  function request_cookie_deletion(doc) {
    let iframe = doc.createElement('iframe');
    iframe.src = "./resources/cookie-setter.py?clean";
    doc.body.appendChild(iframe);
  }

  function create_cookie_test(file) {
    return t => {
      const iframe = document.createElement('iframe');
      document.body.appendChild(iframe);
      // delete_all_cookies(iframe.contentDocument);
      // request_cookie_deletion(iframe.contentDocument);
      iframe.src = "./resources/cookie-setter.py?file="+file;
      // delete_all_cookies(document);
      // request_cookie_deletion(document);
      const kInitialCookies =
          iframe.contentDocument.cookie.replace(/^Cookie: /, '').split(/\s*;\s*/);
      return new Promise((resolve, reject) => {
        // delete_all_cookies(iframe.contentDocument);
        // request_cookie_deletion(iframe.contentDocument);
        window.addEventListener("message", t.step_func(e => {
          if (e.data.uri.match(file) === null) {
            return;  // We intercepted the message of a different call.
          }
          resolve(iframe.contentDocument);
        }));
      }).then((response_doc) => {
        let actual_cookies = response_doc.cookie;
        for (let i in kInitialCookies) {
          let no_spaces_cookie_regex =
              new RegExp(/\s*[\;]*\s/.source + kInitialCookies[i]);
          actual_cookies = actual_cookies.replace(no_spaces_cookie_regex, '');
        }
        let expected_cookies = response_doc.body.innerText
            .replace(/^Cookie: /, '')
            .replace(/^\s+|\s+$/g, '');
        assert_equals(actual_cookies, expected_cookies);
        t.done();
      });
    };
  }

  let cookie_tests = [
    {name: "Set cookie.", file: "0001"}
  ]

  for (let i in cookie_tests) {
    promise_test(create_cookie_test(cookie_tests[i].file),
                 cookie_tests[i].name);
  }
</script>
